requirements.txt
# Core Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Image Processing
opencv-python==4.9.0.80
Pillow==10.3.0

# Data Processing
numpy==1.26.4
pandas==2.2.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Web Application
streamlit==1.37.1

# Utilities
tqdm==4.66.4
scikit-learn==1.4.2
gdown==5.1.0

# Compatibility fix
protobuf==4.25.3

supabase
python-dotenv